backend/requirements.txt
# Web framework
fastapi
uvicorn

# Database
sqlalchemy
pydantic
pydantic[email]

# User authentication
pwdlib[argon2]
pyjwt
pydantic-settings

# File uploads
python-multipart

# Testing
pytest
httpx